requirements.txt
argon2-cffi==20.1.0
attrs==19.3.0
Babel==2.9.1
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
boto3==1.14.28
botocore==1.17.28
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
en-core-web-sm==2.3.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.3.1
flow-utils==0.0.1
FlowUtils==0.9.5
future==0.18.2
h5py==2.10.0
huggingface-hub==0.0.19
idna==2.10
importlib-metadata==1.7.0
ipdb==0.13.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
joypy==0.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.6.3
nltk==3.5
notebook==6.1.1
numpy==1.19.1
openpyxl==3.0.8
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytorch-nlp==0.5.0
pytorch-pretrained-bert==0.6.2
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
sacremoses==0.0.46
scikit-learn==0.23.1
scipy==1.5.2
seaborn==0.11.2
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
spacy==2.3.2
srsly==1.0.2
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.5.1
torchtext==0.6.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.25.10
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
xlsx2html==0.4.0
XlsxWriter==3.0.1
zipp==3.1.0
